backend/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
PyJWT==2.3.0
